third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/getallresponseheaders-expected.txt

This is a testharness.js-based test.
PASS XMLHttpRequest: getAllResponseHeaders()
FAIL XMLHttpRequest: getAllResponseHeaders() 1assert_equals: expected "also-here: Mr. PB\r\newok: lego\r\nfoo-test: 1, 2\r\n" but got "also-here: Mr. PB\r\nfoo-test: 1, 2\r\newok: lego\r\n"
Harness: the test ran to completion.
